requirements.txt
fastapi==0.79.0
orjson==3.4.1
pydantic~=1.10.13
uvicorn==0.18.2
uvloop==0.14.0
SQLAlchemy==1.4.41
asyncpg==0.26.0

python-dotenv~=1.0.0
PyJWT~=2.8.0